Complete UI change.
Completetly removed old UI in triGame.game.playerInterface and replaced it
with triGame.game.ui . The new UI is at the bottom instead of on the
right. It is smaller and easier to understand.

Some upgrade functionality was added to the gun/projectile classes. The buildingclasses are still missing some upgrade functionality.